Extract Segment Repairer out from the Segment Store
