+++ pkg/datarepair/repairer/config.go
@@ -7,18 +7,16 @@ import (
 	"context"
 	"time"
 
-	"github.com/vivint/infectious"
 	"go.uber.org/zap"
 
 	"storj.io/storj/pkg/datarepair/queue"
-	"storj.io/storj/pkg/eestream"
 	"storj.io/storj/pkg/miniogw"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/pointerdb/pdbclient"
 	"storj.io/storj/pkg/provider"
-	ecclient "storj.io/storj/pkg/storage/ec"
-	segment "storj.io/storj/pkg/storage/segments"
+	"storj.io/storj/pkg/storage/ec"
+	"storj.io/storj/pkg/storage/segments"
 	"storj.io/storj/storage/redis"
 )
 
@@ -33,19 +31,6 @@ type Config struct {
 	APIKey        string        `help:"repairer-specific pointerdb access credential"`
 
 	miniogw.NodeSelectionConfig
-
-	// TODO: this is a huge bug that these are required here. these values should
-	// all come from the pointer for each repair. these need to be removed from the
-	// config
-	MinThreshold     int `help:"TODO: remove" default:"29"`
-	RepairThreshold  int `help:"TODO: remove" default:"35"`
-	SuccessThreshold int `help:"TODO: remove" default:"80"`
-	MaxThreshold     int `help:"TODO: remove" default:"95"`
-	ErasureShareSize int `help:"TODO: remove" default:"1024"`
-
-	// TODO: the repairer shouldn't need to worry about inlining, as it is only
-	// repairing non-inlined things.
-	MaxInlineSize int `help:"TODO: remove" default:"4096"`
 }
 
 // Run runs the repairer with configured values
@@ -57,12 +42,12 @@ func (c Config) Run(ctx context.Context, server *provider.Provider) (err error) 
 
 	queue := queue.NewQueue(redisQ)
 
-	ss, err := c.getSegmentStore(ctx, server.Identity())
+	sr, err := c.getSegmentRepairer(ctx, server.Identity())
 	if err != nil {
 		return Error.Wrap(err)
 	}
 
-	repairer := newRepairer(queue, ss, c.Interval, c.MaxRepair)
+	repairer := newRepairer(queue, sr, c.Interval, c.MaxRepair)
 
 	ctx, cancel := context.WithCancel(ctx)
 
@@ -77,8 +62,8 @@ func (c Config) Run(ctx context.Context, server *provider.Provider) (err error) 
 	return server.Run(ctx)
 }
 
-// getSegmentStore creates a new segment store from storeConfig values
-func (c Config) getSegmentStore(ctx context.Context, identity *provider.FullIdentity) (ss segment.Store, err error) {
+// getSegmentRepairer creates a new segment repairer from storeConfig values
+func (c Config) getSegmentRepairer(ctx context.Context, identity *provider.FullIdentity) (ss segments.Repairer, err error) {
 	defer mon.Task()(&ctx)(&err)
 
 	var oc overlay.Client
@@ -93,14 +78,6 @@ func (c Config) getSegmentStore(ctx context.Context, identity *provider.FullIden
 	}
 
 	ec := ecclient.NewClient(identity, c.MaxBufferMem)
-	fc, err := infectious.NewFEC(c.MinThreshold, c.MaxThreshold)
-	if err != nil {
-		return nil, err
-	}
-	rs, err := eestream.NewRedundancyStrategy(eestream.NewRSScheme(fc, c.ErasureShareSize), c.RepairThreshold, c.SuccessThreshold)
-	if err != nil {
-		return nil, err
-	}
 
 	ns := &pb.NodeStats{
 		UptimeRatio:       c.UptimeRatio,
@@ -109,5 +86,5 @@ func (c Config) getSegmentStore(ctx context.Context, identity *provider.FullIden
 		AuditCount:        c.AuditCount,
 	}
 
-	return segment.NewSegmentStore(oc, ec, pdb, rs, c.MaxInlineSize, ns), nil
+	return segments.NewSegmentRepairer(oc, ec, pdb, ns), nil
 }
