--- pkg/datarepair/repairer/repairer_test.go
@@ -1,4 +1,0 @@
-// Copyright (C) 2018 Storj Labs, Inc.
-// See LICENSE for copying information.
-
-package repairer
