+++ pkg/storage/segments/store.go
@@ -21,9 +21,8 @@ import (
 	"storj.io/storj/pkg/piecestore/psclient"
 	"storj.io/storj/pkg/pointerdb/pdbclient"
 	"storj.io/storj/pkg/ranger"
-	ecclient "storj.io/storj/pkg/storage/ec"
+	"storj.io/storj/pkg/storage/ec"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/pkg/utils"
 )
 
 var (
@@ -49,7 +48,6 @@ type ListItem struct {
 type Store interface {
 	Meta(ctx context.Context, path storj.Path) (meta Meta, err error)
 	Get(ctx context.Context, path storj.Path) (rr ranger.Ranger, meta Meta, err error)
-	Repair(ctx context.Context, path storj.Path, lostPieces []int32) (err error)
 	Put(ctx context.Context, data io.Reader, expiration time.Time, segmentInfo func() (storj.Path, []byte, error)) (meta Meta, err error)
 	Delete(ctx context.Context, path storj.Path) (err error)
 	List(ctx context.Context, prefix, startAfter, endBefore storj.Path, recursive bool, limit int, metaFlags uint32) (items []ListItem, more bool, err error)
@@ -66,14 +64,7 @@ type segmentStore struct {
 
 // NewSegmentStore creates a new instance of segmentStore
 func NewSegmentStore(oc overlay.Client, ec ecclient.Client, pdb pdbclient.Client, rs eestream.RedundancyStrategy, threshold int, nodeStats *pb.NodeStats) Store {
-	return &segmentStore{oc: oc, ec: ec, pdb: pdb, rs: rs, thresholdSize: threshold,
-		nodeStats: &pb.NodeStats{
-			UptimeRatio:       nodeStats.UptimeRatio,
-			UptimeCount:       nodeStats.UptimeCount,
-			AuditSuccessRatio: nodeStats.AuditSuccessRatio,
-			AuditCount:        nodeStats.AuditCount,
-		},
-	}
+	return &segmentStore{oc: oc, ec: ec, pdb: pdb, rs: rs, thresholdSize: threshold, nodeStats: nodeStats}
 }
 
 // Meta retrieves the metadata of the segment
@@ -176,40 +167,6 @@ func (s *segmentStore) Put(ctx context.Context, data io.Reader, expiration time.
 	return m, nil
 }
 
-// makeRemotePointer creates a pointer of type remote
-func makeRemotePointer(nodes []*pb.Node, rs eestream.RedundancyStrategy, pieceID psclient.PieceID, readerSize int64, exp *timestamp.Timestamp, metadata []byte) (pointer *pb.Pointer, err error) {
-	var remotePieces []*pb.RemotePiece
-	for i := range nodes {
-		if nodes[i] == nil {
-			continue
-		}
-		remotePieces = append(remotePieces, &pb.RemotePiece{
-			PieceNum: int32(i),
-			NodeId:   nodes[i].Id,
-		})
-	}
-
-	pointer = &pb.Pointer{
-		Type: pb.Pointer_REMOTE,
-		Remote: &pb.RemoteSegment{
-			Redundancy: &pb.RedundancyScheme{
-				Type:             pb.RedundancyScheme_RS,
-				MinReq:           int32(rs.RequiredCount()),
-				Total:            int32(rs.TotalCount()),
-				RepairThreshold:  int32(rs.RepairThreshold()),
-				SuccessThreshold: int32(rs.OptimalThreshold()),
-				ErasureShareSize: int32(rs.ErasureShareSize()),
-			},
-			PieceId:      string(pieceID),
-			RemotePieces: remotePieces,
-		},
-		SegmentSize:    readerSize,
-		ExpirationDate: exp,
-		Metadata:       metadata,
-	}
-	return pointer, nil
-}
-
 // Get retrieves a segment using erasure code, overlay, and pointerdb clients
 func (s *segmentStore) Get(ctx context.Context, path storj.Path) (rr ranger.Ranger, meta Meta, err error) {
 	defer mon.Task()(&ctx)(&err)
@@ -226,7 +183,7 @@ func (s *segmentStore) Get(ctx context.Context, path storj.Path) (rr ranger.Rang
 		seg := pr.GetRemote()
 		pid := psclient.PieceID(seg.GetPieceId())
 
-		nodes, err = s.lookupAndAlignNodes(ctx, nodes, seg)
+		nodes, err = lookupAndAlignNodes(ctx, s.oc, nodes, seg)
 		if err != nil {
 			return nil, Meta{}, Error.Wrap(err)
 		}
@@ -265,6 +222,91 @@ func (s *segmentStore) Get(ctx context.Context, path storj.Path) (rr ranger.Rang
 	return rr, convertMeta(pr), nil
 }
 
+// makeRemotePointer creates a pointer of type remote
+func makeRemotePointer(nodes []*pb.Node, rs eestream.RedundancyStrategy, pieceID psclient.PieceID, readerSize int64, exp *timestamp.Timestamp, metadata []byte) (pointer *pb.Pointer, err error) {
+	var remotePieces []*pb.RemotePiece
+	for i := range nodes {
+		if nodes[i] == nil {
+			continue
+		}
+		remotePieces = append(remotePieces, &pb.RemotePiece{
+			PieceNum: int32(i),
+			NodeId:   nodes[i].Id,
+		})
+	}
+
+	pointer = &pb.Pointer{
+		Type: pb.Pointer_REMOTE,
+		Remote: &pb.RemoteSegment{
+			Redundancy: &pb.RedundancyScheme{
+				Type:             pb.RedundancyScheme_RS,
+				MinReq:           int32(rs.RequiredCount()),
+				Total:            int32(rs.TotalCount()),
+				RepairThreshold:  int32(rs.RepairThreshold()),
+				SuccessThreshold: int32(rs.OptimalThreshold()),
+				ErasureShareSize: int32(rs.ErasureShareSize()),
+			},
+			PieceId:      string(pieceID),
+			RemotePieces: remotePieces,
+		},
+		SegmentSize:    readerSize,
+		ExpirationDate: exp,
+		Metadata:       metadata,
+	}
+	return pointer, nil
+}
+
+// Delete tells piece stores to delete a segment and deletes pointer from pointerdb
+func (s *segmentStore) Delete(ctx context.Context, path storj.Path) (err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	pr, nodes, _, err := s.pdb.Get(ctx, path)
+	if err != nil {
+		return Error.Wrap(err)
+	}
+
+	if pr.GetType() == pb.Pointer_REMOTE {
+		seg := pr.GetRemote()
+		pid := psclient.PieceID(seg.PieceId)
+
+		nodes, err = lookupAndAlignNodes(ctx, s.oc, nodes, seg)
+		if err != nil {
+			return Error.Wrap(err)
+		}
+
+		authorization := s.pdb.SignedMessage()
+		// ecclient sends delete request
+		err = s.ec.Delete(ctx, nodes, pid, authorization)
+		if err != nil {
+			return Error.Wrap(err)
+		}
+	}
+
+	// deletes pointer from pointerdb
+	return s.pdb.Delete(ctx, path)
+}
+
+// List retrieves paths to segments and their metadata stored in the pointerdb
+func (s *segmentStore) List(ctx context.Context, prefix, startAfter, endBefore storj.Path, recursive bool, limit int, metaFlags uint32) (items []ListItem, more bool, err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	pdbItems, more, err := s.pdb.List(ctx, prefix, startAfter, endBefore, recursive, limit, metaFlags)
+	if err != nil {
+		return nil, false, err
+	}
+
+	items = make([]ListItem, len(pdbItems))
+	for i, itm := range pdbItems {
+		items[i] = ListItem{
+			Path:     itm.Path,
+			Meta:     convertMeta(itm.Pointer),
+			IsPrefix: itm.IsPrefix,
+		}
+	}
+
+	return items, more, nil
+}
+
 func makeRedundancyStrategy(scheme *pb.RedundancyScheme) (eestream.RedundancyStrategy, error) {
 	fc, err := infectious.NewFEC(int(scheme.GetMinReq()), int(scheme.GetTotal()))
 	if err != nil {
@@ -296,164 +338,10 @@ func calcNeededNodes(rs *pb.RedundancyScheme) int32 {
 	return needed
 }
 
-// Delete tells piece stores to delete a segment and deletes pointer from pointerdb
-func (s *segmentStore) Delete(ctx context.Context, path storj.Path) (err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	pr, nodes, _, err := s.pdb.Get(ctx, path)
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	if pr.GetType() == pb.Pointer_REMOTE {
-		seg := pr.GetRemote()
-		pid := psclient.PieceID(seg.PieceId)
-
-		nodes, err = s.lookupAndAlignNodes(ctx, nodes, seg)
-		if err != nil {
-			return Error.Wrap(err)
-		}
-
-		authorization := s.pdb.SignedMessage()
-		// ecclient sends delete request
-		err = s.ec.Delete(ctx, nodes, pid, authorization)
-		if err != nil {
-			return Error.Wrap(err)
-		}
-	}
-
-	// deletes pointer from pointerdb
-	return s.pdb.Delete(ctx, path)
-}
-
-// Repair retrieves an at-risk segment and repairs and stores lost pieces on new nodes
-func (s *segmentStore) Repair(ctx context.Context, path storj.Path, lostPieces []int32) (err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	// Read the segment's pointer's info from the PointerDB
-	pr, originalNodes, pba, err := s.pdb.Get(ctx, path)
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	if pr.GetType() != pb.Pointer_REMOTE {
-		return Error.New("cannot repair inline segment %s", psclient.PieceID(pr.GetInlineSegment()))
-	}
-
-	seg := pr.GetRemote()
-	pid := psclient.PieceID(seg.GetPieceId())
-
-	originalNodes, err = s.lookupAndAlignNodes(ctx, originalNodes, seg)
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	// Get the nodes list that needs to be excluded
-	var excludeNodeIDs storj.NodeIDList
-
-	// Count the number of nil nodes thats needs to be repaired
-	totalNilNodes := 0
-
-	healthyNodes := make([]*pb.Node, len(originalNodes))
-
-	// Populate healthyNodes with all nodes from originalNodes except those correlating to indices in lostPieces
-	for i, v := range originalNodes {
-		if v == nil {
-			totalNilNodes++
-			continue
-		}
-
-		excludeNodeIDs = append(excludeNodeIDs, v.Id)
-
-		// If node index exists in lostPieces, skip adding it to healthyNodes
-		if contains(lostPieces, i) {
-			totalNilNodes++
-		} else {
-			healthyNodes[i] = v
-		}
-	}
-
-	// Request Overlay for n-h new storage nodes
-	op := overlay.Options{Amount: totalNilNodes, Space: 0, Excluded: excludeNodeIDs}
-	newNodes, err := s.oc.Choose(ctx, op)
-	if err != nil {
-		return err
-	}
-
-	if totalNilNodes != len(newNodes) {
-		return Error.New("Number of new nodes from overlay (%d) does not equal total nil nodes (%d)", len(newNodes), totalNilNodes)
-	}
-
-	totalRepairCount := len(newNodes)
-
-	// Make a repair nodes list just with new unique ids
-	repairNodes := make([]*pb.Node, len(healthyNodes))
-	for i, vr := range healthyNodes {
-		// Check that totalRepairCount is non-negative
-		if totalRepairCount < 0 {
-			return Error.New("Total repair count (%d) less than zero", totalRepairCount)
-		}
-
-		// Find the nil nodes in the healthyNodes list
-		if vr == nil {
-			// Assign the item in repairNodes list with an item from the newNode list
-			totalRepairCount--
-			repairNodes[i] = newNodes[totalRepairCount]
-		}
-	}
-
-	// Check that all nil nodes have a replacement prepared
-	if totalRepairCount != 0 {
-		return Error.New("Failed to replace all nil nodes (%d). (%d) new nodes not inserted", len(newNodes), totalRepairCount)
-	}
-
-	rs, err := makeRedundancyStrategy(pr.GetRemote().GetRedundancy())
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	signedMessage := s.pdb.SignedMessage()
-
-	// Download the segment using just the healthyNodes
-	rr, err := s.ec.Get(ctx, healthyNodes, rs, pid, pr.GetSegmentSize(), pba, signedMessage)
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	r, err := rr.Range(ctx, 0, rr.Size())
-	if err != nil {
-		return Error.Wrap(err)
-	}
-	defer utils.LogClose(r)
-
-	// Upload the repaired pieces to the repairNodes
-	successfulNodes, err := s.ec.Put(ctx, repairNodes, rs, pid, r, convertTime(pr.GetExpirationDate()), pba, signedMessage)
-	if err != nil {
-		return Error.Wrap(err)
-	}
-
-	// Merge the successful nodes list into the healthy nodes list
-	for i, v := range healthyNodes {
-		if v == nil {
-			// copy the successfuNode info
-			healthyNodes[i] = successfulNodes[i]
-		}
-	}
-
-	metadata := pr.GetMetadata()
-	pointer, err := makeRemotePointer(healthyNodes, rs, pid, rr.Size(), pr.GetExpirationDate(), metadata)
-	if err != nil {
-		return err
-	}
-
-	// update the segment info in the pointerDB
-	return s.pdb.Put(ctx, path, pointer)
-}
-
 // lookupNodes, if necessary, calls Lookup to get node addresses from the overlay.
 // It also realigns the nodes to an indexed list of nodes based on the piece number.
 // Missing pieces are represented by a nil node.
-func (s *segmentStore) lookupAndAlignNodes(ctx context.Context, nodes []*pb.Node, seg *pb.RemoteSegment) (result []*pb.Node, err error) {
+func lookupAndAlignNodes(ctx context.Context, oc overlay.Client, nodes []*pb.Node, seg *pb.RemoteSegment) (result []*pb.Node, err error) {
 	defer mon.Task()(&ctx)(&err)
 
 	if nodes == nil {
@@ -463,7 +351,7 @@ func (s *segmentStore) lookupAndAlignNodes(ctx context.Context, nodes []*pb.Node
 			nodeIds = append(nodeIds, p.NodeId)
 		}
 		// Lookup the node info from node IDs
-		nodes, err = s.oc.BulkLookup(ctx, nodeIds)
+		nodes, err = oc.BulkLookup(ctx, nodeIds)
 		if err != nil {
 			return nil, Error.Wrap(err)
 		}
@@ -476,27 +364,6 @@ func (s *segmentStore) lookupAndAlignNodes(ctx context.Context, nodes []*pb.Node
 	}
 
 	return result, nil
-}
-
-// List retrieves paths to segments and their metadata stored in the pointerdb
-func (s *segmentStore) List(ctx context.Context, prefix, startAfter, endBefore storj.Path, recursive bool, limit int, metaFlags uint32) (items []ListItem, more bool, err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	pdbItems, more, err := s.pdb.List(ctx, prefix, startAfter, endBefore, recursive, limit, metaFlags)
-	if err != nil {
-		return nil, false, err
-	}
-
-	items = make([]ListItem, len(pdbItems))
-	for i, itm := range pdbItems {
-		items[i] = ListItem{
-			Path:     itm.Path,
-			Meta:     convertMeta(itm.Pointer),
-			IsPrefix: itm.IsPrefix,
-		}
-	}
-
-	return items, more, nil
 }
 
 // contains checks if n exists in list
